Cleaned up after useZip hook and time string function
