+++ src/App.js
@@ -5,66 +5,35 @@ import BatchButton from "./components/BatchButton.js";
 import { View, Flex, ActionButton } from "@adobe/react-spectrum";
 import "./App.css";
 import useZip from "./components/useZip.js";
-// import JSZip from "jszip";
-// import FileSaver from "file-saver";
+import GenerateTimeString from "./components/GenerateTimeString.js";
 
 function App() {
   const [progress, setProgress] = useState("hold");
   const [droppedFiles, setDroppedFiles] = useState([]);
   const [filesComplete, setFilesComplete] = useState(0);
+  const generateNewTimeString = GenerateTimeString();
+  const [completeZip, generateNewZip, setZipFolder] = useZip();
 
+  //Trigger processing
   const startTheMagick = () => {
     setProgress("processing");
   };
 
-  const [completeZip, generateNewZip, setZipFolder] = useZip();
-
-  //Get time string if needed for multiple Zip's
-  // const intlOptions = {
-  //   hour: "2-digit",
-  //   minute: "2-digit",
-  //   second: "2-digit",
-  //   hour12: false,
-  // };
-  //
-  // const generateNewTimeString = () => {
-  //   return new Intl.DateTimeFormat("default", intlOptions)
-  //     .format(new Date())
-  //     .replace(/:/g, ".");
-  // };
-
-  //Setup Zip
-  // const [zip, setZip] = useState({});
-  // const [, setZipFolder] = useState({});
-
   useEffect(() => {
-    console.log(generateNewZip);
-    // let newZip = new JSZip();
-    // let newFolder = newZip.folder(generateNewTimeString());
-    //
-    // setZip(newZip);
-    // setZipFolder(newFolder);
     //Generate initial Zip
-    generateNewZip();
+    generateNewZip(generateNewTimeString());
+    //eslint-disable-next-line
   }, []);
-  //
-  // const completeZip = () => {
-  //   zip.generateAsync({ type: "blob" }).then(blob => {
-  //     FileSaver.saveAs(blob, "shrunk.zip");
-  //   });
-  // };
 
   const resetApp = () => {
-    // let newZip = new JSZip();
-    // let newFolder = newZip.folder(generateNewTimeString());
-
-    generateNewZip();
+    generateNewZip(generateNewTimeString());
     setDroppedFiles([]);
     setFilesComplete(0);
     setProgress("hold");
   };
 
   useEffect(() => {
+    //Watch total files vs 'completed' and set to complete if equal
     if (droppedFiles.length > 0 && filesComplete === droppedFiles.length) {
       //Set completed to NaN to stop infinite re-renders for matching if statement
       setFilesComplete(NaN);
@@ -88,9 +57,8 @@ function App() {
       },
       false
     );
-
+    //Cleanup listener on component unmounting
     return () => {
-      //Cleanup listener on component unmounting
       window.removeEventListener("dragover", e => {
         e.preventDefault();
       });
